pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "miao"
version = "0.1.0"
description = "Scalable PyTorch data loaders for OME-NGFF zarr datasets"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "tensorstore>=0.1.60",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "zarr>=2.16",
]

[tool.hatch.build.targets.wheel]
packages = ["src/miao"]
